address xinrong comment and move to pandas.utils
