+++ python/pyspark/sql/connect/udf.py
@@ -36,10 +36,8 @@ from pyspark.sql.connect.expressions import (
 from pyspark.sql.connect.column import Column
 from pyspark.sql.connect.types import UnparsedDataType
 from pyspark.sql.types import ArrayType, DataType, MapType, StringType, StructType
-from pyspark.sql.udf import (
-    UDFRegistration as PySparkUDFRegistration,
-    _is_barrier,
-)
+from pyspark.sql.udf import UDFRegistration as PySparkUDFRegistration
+from pyspark.sql.pandas.utils import _is_barrier
 from pyspark.errors import PySparkTypeError
