+++ python/pyspark/sql/pandas/functions.py
@@ -22,9 +22,13 @@ from typing import get_type_hints
 
 from pyspark.rdd import PythonEvalType
 from pyspark.sql.pandas.typehints import infer_eval_type
-from pyspark.sql.pandas.utils import require_minimum_pandas_version, require_minimum_pyarrow_version
+from pyspark.sql.pandas.utils import (
+    require_minimum_pandas_version,
+    require_minimum_pyarrow_version,
+    _is_barrier,
+)
 from pyspark.sql.types import DataType
-from pyspark.sql.udf import _create_udf, _is_barrier
+from pyspark.sql.udf import _create_udf
 from pyspark.sql.utils import is_remote
 from pyspark.errors import PySparkTypeError, PySparkValueError
