+++ python/pyspark/sql/pandas/utils.py
@@ -16,6 +16,9 @@
 #
 
 
+from typing import Any, Callable
+
+
 def require_minimum_pandas_version() -> None:
     """Raise ImportError if minimum version of Pandas is not installed"""
     # TODO(HyukjinKwon): Relocate and deduplicate the version specification.
@@ -84,3 +87,57 @@ def pyarrow_version_less_than_minimum(minimum_pyarrow_version: str) -> bool:
         return False
 
     return LooseVersion(pyarrow.__version__) < LooseVersion(minimum_pyarrow_version)
+
+
+def barrier(f: Callable) -> Callable:
+    """Mark functon should be executed in barrier mode. This API is a developer API.
+
+    .. versionadded:: 3.5.0
+
+    Parameters
+    ----------
+    f : function, :meth:`pyspark.sql.functions.udf` or :meth:`pyspark.sql.functions.pandas_udf`
+        a Python function, or a user-defined function. The user-defined function can
+        be either row-at-a-time or vectorized. See :meth:`pyspark.sql.functions.udf` and
+        :meth:`pyspark.sql.functions.pandas_udf`.
+
+    Returns
+    -------
+    function
+        a user-defined function
+
+    Notes
+    -----
+    The barrier function is dedicated for external ML frameworks, including PyTorch and XGBoost.
+    It is supposed only to be used in mapInPandas and mapInArrow, and followed by a collection to
+    get the model coefficients. Also note that some DataFrame operations (e.g. df.show, df.take)
+    are not allowed in barrier mode, due to the underlying RDD operations are not supported.
+
+    Examples
+    --------
+    >>> from pyspark.sql.pandas.utils import barrier
+
+    create a dataframe
+    >>> df = spark.createDataFrame([(1, 21), (2, 30)], ("id", "age"))
+
+    define a barrier python udf
+    >>> @barrier
+    ... def filter_func(iterator):
+    ...     for pdf in iterator:
+    ...         yield pdf[pdf.id == 1]
+    ...
+    >>> df.mapInPandas(filter_func, df.schema).collect()
+    [Row(id=1, age=21)]
+    """
+
+    f._is_barrier = True  # type: ignore[attr-defined]
+    return f
+
+
+def _is_barrier(obj: Any) -> bool:
+    return (
+        obj is not None
+        and hasattr(obj, "_is_barrier")
+        and isinstance(obj._is_barrier, bool)
+        and obj._is_barrier
+    )
