+++ python/pyspark/sql/udf.py
@@ -42,7 +42,11 @@ from pyspark.sql.types import (
 )
 from pyspark.sql.utils import get_active_spark_context
 from pyspark.sql.pandas.types import to_arrow_type
-from pyspark.sql.pandas.utils import require_minimum_pandas_version, require_minimum_pyarrow_version
+from pyspark.sql.pandas.utils import (
+    require_minimum_pandas_version,
+    require_minimum_pyarrow_version,
+    _is_barrier,
+)
 from pyspark.errors import PySparkTypeError, PySparkNotImplementedError
 
 if TYPE_CHECKING:
@@ -50,61 +54,6 @@ if TYPE_CHECKING:
     from pyspark.sql.session import SparkSession
 
 __all__ = ["UDFRegistration"]
-
-
-def _is_barrier(obj: Any) -> bool:
-    return (
-        obj is not None
-        and hasattr(obj, "_is_barrier")
-        and isinstance(obj._is_barrier, bool)
-        and obj._is_barrier
-    )
-
-
-def barrier(f: Callable) -> Callable:
-    """Mark functon should be executed in barrier mode. This API is a developer API.
-
-    .. versionadded:: 3.5.0
-
-    Parameters
-    ----------
-    f : function, :meth:`pyspark.sql.functions.udf` or :meth:`pyspark.sql.functions.pandas_udf`
-        a Python function, or a user-defined function. The user-defined function can
-        be either row-at-a-time or vectorized. See :meth:`pyspark.sql.functions.udf` and
-        :meth:`pyspark.sql.functions.pandas_udf`.
-
-    Returns
-    -------
-    function
-        a user-defined function
-
-    Notes
-    -----
-    The barrier function is dedicated for external ML frameworks, including PyTorch and XGBoost.
-    It is supposed only to be used in mapInPandas and mapInArrow, and followed by a collection to
-    get the model coefficients. Also note that some DataFrame operations (e.g. df.show, df.take)
-    are not allowed in barrier mode, due to the underlying RDD operations are not supported.
-
-    Examples
-    --------
-    >>> from pyspark.sql.udf import barrier
-
-    create a dataframe
-    >>> df = spark.createDataFrame([(1, 21), (2, 30)], ("id", "age"))
-
-    define a barrier python udf
-    >>> @barrier
-    ... def filter_func(iterator):
-    ...     for pdf in iterator:
-    ...         yield pdf[pdf.id == 1]
-    ...
-    >>> df.mapInPandas(filter_func, df.schema).collect()
-    [Row(id=1, age=21)]
-    """
-
-    wrapped = f
-    wrapped._is_barrier = True  # type: ignore[attr-defined]
-    return wrapped
 
 
 def _wrap_function(
